requirements.txt
python-telegram-bot
requests
openai
python-dotenv
